+++ backend/packages/fiab-plugin-ecmwf/src/fiab_plugin_ecmwf/output/__init__.py
@@ -6,7 +6,3 @@
 # In applying this licence, ECMWF does not waive the privileges and immunities
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
-
-from .blocks import GribSink, ZarrSink
-
-__all__ = ["ZarrSink", "GribSink"]
